requirements.txt
discord
openai== 1
python-dotenv